finnaly add calculation SAW

+++ src/com/main/models/entity/dataBobotKriteria.java
@@ -3,9 +3,9 @@ package com.main.models.entity;
 public class dataBobotKriteria {
     private int idKriteria;
     private String kriteria, type;
-    private double weight;
+    private int weight;
 
-    public dataBobotKriteria(int idKriteria, String kriteria, double weight, String type) {
+    public dataBobotKriteria(int idKriteria, String kriteria, int weight, String type) {
         this.idKriteria = idKriteria;
         this.kriteria = kriteria;
         this.weight = weight;
@@ -20,7 +20,7 @@ public class dataBobotKriteria {
         return kriteria;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
